More Actum Daysim to current Daysim conversions
Compared Pre-Peter current DaySim with John's Daysim and tried to bring all new John code into Actum specific Domain Models

+++ DaySim.Framework/DomainModels/Models/IPerson.cs
@@ -37,19 +37,19 @@ namespace DaySim.Framework.DomainModels.Models {
 
     double AutoDistanceToUsualSchool { get; set; }
 
-    int UsualModeToWork { get; set; }
+    double UsualModeToWork { get; set; }
 
-    int UsualArrivalPeriodToWork { get; set; }
+    double UsualArrivalPeriodToWork { get; set; }
 
-    int UsualDeparturePeriodFromWork { get; set; }
+    double UsualDeparturePeriodFromWork { get; set; }
 
     int TransitPassOwnership { get; set; }
 
     int PaidParkingAtWorkplace { get; set; }
 
-    int PaperDiary { get; set; }
+    double PaperDiary { get; set; }
 
-    int ProxyResponse { get; set; }
+    double ProxyResponse { get; set; }
 
     double ExpansionFactor { get; set; }
   }
